+++ django-rgd/client/rgd_client/client.py
@@ -1,18 +1,12 @@
 import getpass
-import inspect
 import os
-from typing import Dict, List, Optional, Type
+from typing import List, Optional, Type
 
-from pkg_resources import iter_entry_points
 import requests
 
-from .plugin import CorePlugin, RGDPlugin
+from .plugin import CorePlugin
 from .session import RgdClientSession, clone_session
 from .utils import API_KEY_DIR_PATH, API_KEY_FILE_NAME, DEFAULT_RGD_API
-
-_NAMESPACE = 'rgd_client.plugin'
-_PLUGIN_CLASS_DICT = Dict[str, Type[RGDPlugin]]
-_PLUGIN_INSTANCE_DICT = Dict[Type[RGDPlugin], RGDPlugin]
 
 
 class RgdClient:
@@ -55,58 +49,6 @@ class RgdClient:
         (API_KEY_DIR_PATH / API_KEY_FILE_NAME).unlink(missing_ok=True)
 
 
-def _plugin_classes(extra_plugins: Optional[List] = None) -> _PLUGIN_CLASS_DICT:
-    """Return a dict that maps a plugin namespace to its class."""
-    entry_points = iter_entry_points(_NAMESPACE)
-    plugins_classes = [ep.load() for ep in entry_points]
-    if extra_plugins is not None:
-        plugins_classes.extend(extra_plugins)
-
-    members = {}
-    for cls in plugins_classes:
-        members.update(
-            {
-                n: v
-                for n, v in inspect.getmembers(cls)
-                if inspect.isclass(v) and issubclass(v, RGDPlugin)
-            }
-        )
-
-    return members
-
-
-def _plugin_instances(
-    client: RgdClient, plugin_classes: _PLUGIN_CLASS_DICT
-) -> _PLUGIN_INSTANCE_DICT:
-    """Return a dict that maps a plugin class to its instance."""
-    instance_dict: _PLUGIN_INSTANCE_DICT = {CorePlugin: client.rgd}
-    for name, cls in plugin_classes.items():
-        instance = cls(clone_session(client.session))
-        setattr(client, name, instance)
-        instance_dict[cls] = instance
-
-    return instance_dict
-
-
-def _inject_plugin_deps(plugin_instances: _PLUGIN_INSTANCE_DICT):
-    """Inject plugin dependencies for each plugin instance."""
-    for plugin_class, plugin_instance in plugin_instances.items():
-        # Ensure plugins class is defined
-        if not inspect.isclass(getattr(plugin_class, 'plugins', None)):
-            continue
-
-        # Retrieve deps
-        deps = [
-            (name, val)
-            for name, val in inspect.getmembers(plugin_class.plugins)
-            if inspect.isclass(val) and issubclass(val, RGDPlugin)
-        ]
-
-        for name, cls in deps:
-            if cls in plugin_instances:
-                setattr(plugin_instance.plugins, name, plugin_instances[cls])
-
-
 def _get_api_key(api_url: str, username: str, password: str, save: bool) -> str:
     """Get an RGD API Key for the given user from the server, and save it if requested."""
     resp = requests.post(f'{api_url}/api-token-auth', {'username': username, 'password': password})
@@ -145,6 +87,9 @@ def create_rgd_client(
     save: Optional[bool] = True,
     extra_plugins: Optional[List[Type]] = None,
 ):
+    # Avoid circular import
+    from ._plugin_utils import _plugin_classes, _plugin_instances, _inject_plugin_deps
+
     # Create initial client
     client = RgdClient(api_url, username, password, save)
